+++ util/string.ts
@@ -61,7 +61,7 @@ export const print = (childNodes: readonly AstNodes[], opt: PrintOpt = {}): stri
 
 /* NOT FOR BROWSER */
 
-export const encode = factory(/[<>[\]#|=]+/gu, p => encodeURIComponent(p));
+export const encode = factory(/[<>[\]#|=]+/gu, encodeURIComponent);
 
 /**
  * convert newline in text nodes to single whitespace