+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Parse the `caption` attribute of a `<gallery>` extension tag
 - Fix [`LinkToken.prototype.pipeTrick`](https://github.com/bhsd-harry/wikiparser-node/wiki/LinkToken#pipetrick)
 - [`Token.prototype.solveConst`](https://github.com/bhsd-harry/wikiparser-node/wiki/Token#solveconst) when the default value of a [`ArgToken`](https://github.com/bhsd-harry/wikiparser-node/wiki/ArgToken) contains `Token`
+- Absolute path on Windows in [`Parser.config`](https://github.com/bhsd-harry/wikiparser-node/wiki/Parser#config) and [`Parser.i18n`](https://github.com/bhsd-harry/wikiparser-node/wiki/Parser#i18n)
 
 **Changed**